pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-skill-hub"
version = "0.1.0"
description = "Local MCP server for semantic skill search with usage-based learning"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Carlo Cancellieri" },
]
keywords = ["mcp", "claude", "skills", "semantic-search", "ollama"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "watchdog>=4.0.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "jinja2>=3.1",
    "python-multipart>=0.0.9",
    "aiofiles>=23",
    "psutil>=5.9",
    "sqlite-vec>=0.1.6",
    "litellm>=1.52",
    "pyyaml>=6.0",
]

[project.urls]
Repository = "https://github.com/ccancellieri/mcp-skill-hub"

[project.scripts]
skill-hub = "skill_hub.server:main"
skill-hub-cli = "skill_hub.cli:main"
skill-hub-repl = "skill_hub.repl:main"
skill-hub-dashboard = "skill_hub.webapp.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/skill_hub"]

[tool.hatch.envs.default]
type = "virtual"
path = ".venv"
